QT: warning removals

+++ src/qt/widgets/properties/stringmultilinepropertywidgetqt.cpp
@@ -150,7 +150,7 @@ void MultilineTextEdit::adjustHeight() {
         glm::clamp(static_cast<int>(document()->size().height()), minLineCount_, maxLineCount_) *
         lineHeight_;
     QMargins margins(contentsMargins());
-    setFixedHeight(textHeight + margins.top() + margins.bottom());
+    setFixedHeight(static_cast<int>(textHeight + margins.top() + margins.bottom()));
 }
 
 }  // namespace
